Ported to Android 4.4.4_r2
Ported current version of EFS to android-4.4.4_r2.
Also updated the apply_patches.sh script in order to try and
find the current tree version and apply the correct patches.

+++ integration/apply_patches.sh
@@ -1,17 +1,25 @@
 #!/bin/bash
 
+basepath="external/efs-tools/integration/user"
+version=$(repo info Manifest | sed -n 2p | cut -f4 -d ' ')
+
+if [ ! -d "$basepath/$version" ]; then
+    echo "Unsupported version"
+    exit 1
+fi
+
 # force a clean tree before applying patches
 repo forall -c git reset --hard HEAD
 repo forall -c git clean -df
 # Vold integration
-cp external/efs-tools/integration/vold/EncryptedFileStorageCmd.cpp system/vold/
-cp external/efs-tools/integration/vold/EncryptedFileStorageCmd.h system/vold/
-git apply external/efs-tools/integration/vold/Vold-integration.patch --directory=system/vold/
+cp "$basepath/$version/vold/EncryptedFileStorageCmd.cpp" system/vold/
+cp "$basepath/$version/vold/EncryptedFileStorageCmd.h" system/vold/
+git apply "$basepath/$version/vold/Vold-integration.patch" --directory=system/vold/
 # Core integration
-git apply external/efs-tools/integration/core/Core-integration.patch --directory=system/core/
+git apply "$basepath/$version/core/Core-integration.patch" --directory=system/core/
 # Framework integration
-git apply external/efs-tools/integration/frameworks-base/MountManagerService-integration.patch --directory=frameworks/base/
+git apply "$basepath/$version/frameworks-base/MountManagerService-integration.patch" --directory=frameworks/base/
 # Settings integration
-git apply external/efs-tools/integration/settings/Settings-integration.patch --directory=packages/apps/Settings/
+git apply "$basepath/$version/settings/Settings-integration.patch" --directory=packages/apps/Settings/
 # SeLinux policy integration
-git apply external/efs-tools/integration/sepolicy/EFS-SeLinux-Policy.patch --directory=external/sepolicy/
+git apply "$basepath/$version/sepolicy/EFS-SeLinux-Policy.patch" --directory=external/sepolicy/
